requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0

# Groq for LLM
groq>=0.4.0

# Data processing and visualization
pandas>=2.0.0
plotly>=5.17.0

# Additional utilities
pydantic>=2.0.0
aiosqlite>=0.19.0